Create Ventas_por_branch_filter_joins
This query calculates the total inventory cost for each branch in department '08' by joining item, price, stock, and branch information from the bstore schema. The final output shows the branchDescription and the corresponding sum of the inventory cost.

Its made this way so it can be filtered by Item and by branch with metabase filters.